+++ src/test/java/com/caoccao/javet/tutorial/DecimalJavetInV8Mode.java
@@ -77,7 +77,7 @@ public class DecimalJavetInV8Mode implements IJavetClosable {
                         "a.add(b).toString();").executeString());
         try (V8ValueFunction v8ValueFunctionDecimal = v8Runtime.getGlobalObject().get("Decimal")) {
             try (V8ValueObject v8ValueObjectDecimal = v8ValueFunctionDecimal.callAsConstructor(
-                    new V8ValueString("123.45"))) {
+                    v8Runtime.createV8ValueString("123.45"))) {
                 getLogger().logInfo(v8ValueObjectDecimal.toString());
                 if (v8ValueObjectDecimal.has("constructor")) {
                     try (V8ValueFunction v8ValueFunction = v8ValueObjectDecimal.get("constructor")) {