+++ src/test/java/com/caoccao/javet/values/reference/TestV8ValueObject.java
@@ -132,7 +132,7 @@ public class TestV8ValueObject extends BaseTestJavetRuntime {
     public void testGetSetDelete() throws JavetException {
         try (V8ValueObject v8ValueObject = v8Runtime.getExecutor("const a = {}; a;").execute()) {
             assertTrue(v8ValueObject.set("a", v8Runtime.createV8ValueInteger(1)));
-            assertTrue(v8ValueObject.set("b", new V8ValueString("2")));
+            assertTrue(v8ValueObject.set("b", v8Runtime.createV8ValueString("2")));
             assertEquals(1, v8ValueObject.getInteger("a"));
             assertEquals("2", v8ValueObject.getString("b"));
             assertTrue(v8ValueObject.delete("x"));
@@ -204,7 +204,7 @@ public class TestV8ValueObject extends BaseTestJavetRuntime {
         try (V8ValueObject outerObject = v8Runtime.getExecutor("const o = {}; o;").execute()) {
             assertEquals("{}", outerObject.toJsonString());
             try (V8ValueObject innerObject = v8Runtime.createV8ValueObject()) {
-                innerObject.set("a", new V8ValueString("1"));
+                innerObject.set("a", v8Runtime.createV8ValueString("1"));
                 outerObject.set("x", innerObject);
             }
             assertEquals("{\"x\":{\"a\":\"1\"}}", outerObject.toJsonString());
@@ -220,12 +220,12 @@ public class TestV8ValueObject extends BaseTestJavetRuntime {
             try (IV8ValueArray iV8ValueArray = v8ValueObject.getOwnPropertyNames()) {
                 assertEquals(0, iV8ValueArray.getLength());
             }
-            v8ValueObject.setProperty("a", new V8ValueString("1"));
-            v8ValueObject.setProperty(new V8ValueString("b"), v8Runtime.createV8ValueInteger(2));
-            v8ValueObject.setProperty(new V8ValueString("c"), v8Runtime.createV8ValueLong(3));
-            v8ValueObject.setProperty(new V8ValueString("d"), new V8ValueZonedDateTime(now));
-            v8ValueObject.setProperty(new V8ValueString("e"), new V8ValueDouble(1.23));
-            v8ValueObject.setProperty(new V8ValueString("f"), v8Runtime.createV8ValueBoolean(true));
+            v8ValueObject.setProperty("a", v8Runtime.createV8ValueString("1"));
+            v8ValueObject.setProperty(v8Runtime.createV8ValueString("b"), v8Runtime.createV8ValueInteger(2));
+            v8ValueObject.setProperty(v8Runtime.createV8ValueString("c"), v8Runtime.createV8ValueLong(3));
+            v8ValueObject.setProperty(v8Runtime.createV8ValueString("d"), new V8ValueZonedDateTime(now));
+            v8ValueObject.setProperty(v8Runtime.createV8ValueString("e"), new V8ValueDouble(1.23));
+            v8ValueObject.setProperty(v8Runtime.createV8ValueString("f"), v8Runtime.createV8ValueBoolean(true));
             try (IV8ValueArray iV8ValueArray = v8ValueObject.getOwnPropertyNames()) {
                 assertEquals(6, iV8ValueArray.getLength());
                 assertEquals("a", iV8ValueArray.getString(0));
@@ -251,7 +251,7 @@ public class TestV8ValueObject extends BaseTestJavetRuntime {
     @Test
     public void testToClone() throws JavetException {
         try (V8ValueObject v8ValueObject = v8Runtime.getExecutor("const x = {}; x;").execute()) {
-            v8ValueObject.setProperty("a", new V8ValueString("1"));
+            v8ValueObject.setProperty("a", v8Runtime.createV8ValueString("1"));
             assertEquals("{\"a\":\"1\"}", v8ValueObject.toJsonString());
             try (V8ValueObject clonedV8ValueObject = v8ValueObject.toClone()) {
                 assertEquals("{\"a\":\"1\"}", clonedV8ValueObject.toJsonString());
@@ -264,7 +264,7 @@ public class TestV8ValueObject extends BaseTestJavetRuntime {
     @Test
     public void testToJsonString() throws JavetException {
         try (V8ValueObject v8ValueObject = v8Runtime.getExecutor("const x = {}; x;").execute()) {
-            v8ValueObject.setProperty("a", new V8ValueString("1"));
+            v8ValueObject.setProperty("a", v8Runtime.createV8ValueString("1"));
             v8ValueObject.setProperty("b", v8Runtime.createV8ValueInteger(2));
             v8ValueObject.setProperty("c", new V8ValueDouble(1.23));
             assertEquals("{\"a\":\"1\",\"b\":2,\"c\":1.23}", v8ValueObject.toJsonString());