+++ docs/release_notes.rst
@@ -2,6 +2,13 @@
 Release Notes
 =============
 
+0.8.5
+-----
+
+* Added ``createV8ValueString()`` to ``V8Runtime``
+* Added ``JavetError`` with error codes
+* Removed a few exceptions
+
 0.8.4
 -----